Add exercises 4 & 5 and a disclaimer to 2&3
Add exercises 4 and 5 where the knowledge from previous exercises is
applied. In exercise 4, the student must switch from NGINX to Apache as
a web server. In exercise 5, the usage of an existing Dockerfile is
demonstrated for building a site in OpenShift.

Also, since the method we use for adding index.html to the site in
exercises 2 and 3 is not something you would want to do on a real site,
add disclaimers to these exercises explaining this and mention what
users should do instead.