Add a top level README.md for exercises
Add an introductory README.md file to read before starting any
exercises. This file instructs you to clone the Git repo first to make
things easier and gives some generic instructions about completing the
exercises.

Also amend exercise 0 based on some of the instructions in the new
README file: add a comment about checking login status and the version
of Kubernetes/OpenShift so that the student knows to use the correct
version of documentation for Kubernetes/OpenShift.